Refactor ConnectionDetails Tabs
